# goserver/api/openapi.yaml
openapi: 3.0.0
info:
  title: WebRTC Signaling Server
  description: Python WebRTC signaling server with WebSocket support
  version: 1.0.0
paths:
...
  /api/v1/users:
    get:
      tags: [User]
      summary: List all registered users on the server
      description: |
        Returns every user account (directory / address book). Optional `avatar_url` is present when the user has uploaded an avatar.

        Requires a valid `Authorization: Bearer <access_token>` header.
      operationId: listServerUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar_url:
          type: string
          description: Present when the user has an avatar; relative path e.g. /api/v1/avatars/{username}
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum:
            - INVALID_CREDENTIALS
            - TOKEN_EXPIRED
            - INVALID_REFRESH_TOKEN
            - UNAUTHORIZED
            - VALIDATION_ERROR
            - NOT_FOUND
            - TOO_MANY_REQUESTS
            - SERVER_ERROR
        message:
          type: string
      required: [code, message]
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: "UNAUTHORIZED"
                message: "Missing or invalid access token."
...
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              value:
                code: "SERVER_ERROR"
                message: "An unexpected error occurred."