# vueclient/src/api/models/UserProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type UserProfile = {
    id?: string;
    username?: string;
    avatar_url?: string | null;
};